Add access request label to extensible repo settings for customer repos
This label should be used to request access to Ingeno's private packages
hosted on GitHub.

The label was created manually as a default repository label in the
`ingeno` GitHub organization and will be available on repositories not
using the `settings` app, but those repositories using the settings app
and the `customer-template` would override the default labels without
the change in this commit.

See: https://docs.github.com/en/organizations/managing-organization-settings/managing-default-labels-for-repositories-in-your-organization