Add markdown issue template for access requests
This commit builds on d44d396 and adds a Markdown issue template,
because form templates are not supported for private repositories at the
moment.

The `access-request` label needs to exist on the repository to be
applied at issue creation. This was done at the organization level and
templates have been modified as well.

See: https://github.com/github/feedback/discussions/4264