Fix default assignee for access request issue template
Tests have shown that not prefixing the team name with the organization
name results in an invalid assignee name. No default assignee was
assigned when creating an issue from this template.

Although only the Markdown version of the template currently works, this
commit fixes the default assignee on the issue form template as well to
avoid discrepancies between the two versions of the same template
concept.